# Client-side chunks 4: integrations