--- a/crates/re_sdk/Cargo.toml
+++ b/crates/re_sdk/Cargo.toml
@@ -51,6 +51,7 @@ web_viewer = [
 
 [dependencies]
 re_build_info.workspace = true
+re_chunk.workspace = true
 re_log_encoding = { workspace = true, features = ["encoder"] }
 re_log_types.workspace = true
 re_log.workspace = true