--- a/crates/re_sdk/src/lib.rs
+++ b/crates/re_sdk/src/lib.rs
@@ -78,10 +78,7 @@ pub mod log {
         Chunk, ChunkBatcher, ChunkBatcherConfig, ChunkBatcherError, ChunkBatcherResult, ChunkError,
         ChunkResult, PendingRow, TransportChunk,
     };
-    pub use re_log_types::{
-        DataCell, DataRow, DataTable, DataTableBatcher, DataTableBatcherConfig, LogMsg, RowId,
-        TableId,
-    };
+    pub use re_log_types::{DataCell, DataRow, DataTable, LogMsg, RowId, TableId};
 }
 
 /// Time-related types.