--- a/rerun_py/Cargo.toml
+++ b/rerun_py/Cargo.toml
@@ -41,6 +41,7 @@ web_viewer = [
 
 [dependencies]
 re_build_info.workspace = true
+re_chunk.workspace = true
 re_log = { workspace = true, features = ["setup"] }
 re_log_types.workspace = true
 re_memory.workspace = true